now you can shoot to asteroids and collisions are actually working

+++ src/core/appConfig.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <string>
 struct AppConfig {
-    size_t Width = 800;
-    size_t Height = 600;
+    float Width = 800;
+    float Height = 600;
     std::string Title = "Application";
 };
